% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCovars.R
\name{addCovars}
\alias{addCovars}
\title{Add covariates to flight data}
\usage{
addCovars(flightgeo, sigmetgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}

\item{sigmetgeo}{A df with rows as sigmets, and columns as time, location, etc}
}
\value{
A df based on flightgeo, with extra columns for all covariates.
}
\description{
Given basic parsed flight geo data, add columns including "route",
"country", "continent", sigmet overlaps, weather station density,
U.S. storm event overlap and U.S. delay data for use in analyses and 
MonaRk shiny app.
}
