% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareFlights.R
\name{addDistanceCovars}
\alias{addDistanceCovars}
\title{Find shortest distance and duration for routes}
\usage{
addDistanceCovars(flightgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}
}
\value{
A sf dataframe with additional columns containing routeCount, shortest dist and dur
}
\description{
Loop through the routes in flightgeo, find the
existing flight path with the shortest duration, shortest distance("shortestDistFlight"),
as well as the direct geo distance between airports ("shortestDist"), the number of 
flights within this route ("routeCount"). Also returns starting and ending airports 
latitude for data quality analyses.
This function needs to be run after parsing flight data and adding covars,
called within compareFlights() for a shorter pipeline.
}
