% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatesParsing.R
\name{addMondays}
\alias{addMondays}
\title{Save flightgeo into 2 RDS's}
\usage{
addMondays(appDir, statesDataDir, routeCountCutoff = 50)
}
\arguments{
\item{appDir}{Where the MonaRk package files are stored (including "MonaRk")}

\item{statesDataDir}{The location where the large flight state data is downloaded and stored}

\item{routeCountCutoff}{Min number of routes to be included for app}

\item{flightgeo}{A dataframe of all the flights information within routes of interest including geo data.}

\item{geoname}{The name of the geo RDS}

\item{dfname}{The name of the flight data RDS}

\item{folder}{The location where geo and flight data should be saved}

\item{geoSaveDir}{The location of the geo.RDS for each date}

\item{dfSaveDir}{The location of df.RDS for each date}
}
\value{
No object returned, only data written

Add multiple Mondays worth of data

Run this function if it has been several weeks since data has been updated.
It will update delay, storm, sigmet, and flight data.

No object returned, only data written
}
\description{
After making updates to flightgeo, save as two separate RDS's.
THIS FUNCTION PROB ISN'T USEFUL!!!
}
