% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayParsing.R
\name{cleanDelay}
\alias{cleanDelay}
\title{Parse delay data into usable format}
\usage{
cleanDelay(folder)
}
\arguments{
\item{folder}{Location of delay CSVs}
}
\value{
A dataframe containing flight delay information
}
\description{
Delay data is downloaded from https://www.transtats.bts.gov/Tables.asp?DB_ID=120.
Keep all delay CSVs in a single folder and pass it to this function to create a
df to be used in further flight data processing
}
