% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StormsParsing.R
\name{cleanStorms}
\alias{cleanStorms}
\title{Parse Storm Event data into usable format}
\usage{
cleanStorms(folder)
}
\arguments{
\item{folder}{Location of storm CSVs}
}
\value{
A dataframe containing Storm Event information
}
\description{
Storm event data is downloaded from https://www1.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/
Keep all storm event CSVs in a single folder and pass it to this function to create a
df to be used in further flight data processing. Filters out storm events that 
have no relevance to flights. Converts all timezone data to UTC.
}
