% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareFlights.R
\name{compareFlights}
\alias{compareFlights}
\title{Calculate extra distance and duration for flights}
\usage{
compareFlights(flightgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}
}
\value{
A sf dataframe with additional columns containing extra distance and duration
}
\description{
After calculating the distance and duration for each flight, pass to addDistanceCovars to find
the shortest distance and duration for a route and the direct distance of the route, then
calculate the diffence to find "extra" distance and duration
}
