% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppUtils.R
\name{comprss}
\alias{comprss}
\title{Compress numeric values to redable dollar values}
\usage{
comprss(tx)
}
\arguments{
\item{tx}{a number}
}
\value{
A string with a rounded number and a suffix describing its magnitude
}
\description{
Round numbers and display digits as human-readable terms like "K" for thousand,
"M" for million, and so on.
}
