% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCovars.R
\name{countStations}
\alias{countStations}
\title{Count weather stations per KM^2}
\usage{
countStations(flightgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}
}
\value{
A df of weather station densities, with rows corresponding to flights in flightgeo
}
\description{
Read in 3 types of weather station locations. Then for each flight, loop through 
to find the midpoint between the two airports, and transform the flight data to the 
UTM zone of the midpoint. Draw a 230km buffer around the entire flight and calculate
the area of this buffer region in km^2. Then count the number of each station type 
within the buffer to calculate the station density.
}
