% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareFlights.R
\name{createRouteFolders}
\alias{createRouteFolders}
\title{Format route data to be read into MonaRk shiny app}
\usage{
createRouteFolders(flightgeo, geomatrix, appDataDir)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}

\item{geomatrix}{A list of matrices containing 10-second data for lat, long, alt, velocity; "geo.RDS"}

\item{appDataDir}{The folder where all the route subfolders should be located}
}
\value{
No objects returns. Folders are created locally and data is written to them.
}
\description{
All the flight data is too big to be read into the shiny app. Therefore, the app
only reads in data for routes the user selected. This function creates the folder structure
so each route has a folder containing a df.RDS and a geo.RDS that can be read into the app.
}
