% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareFlights.R
\name{createRouteLevelData}
\alias{createRouteLevelData}
\title{Summarize route data}
\usage{
createRouteLevelData(flightgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}
}
\value{
A df summarizing each route's data
}
\description{
Create a df where each row is a route and columns contain summary data.
Used to compare routes and select routes of interest.
This returned df is shown on the first tab of the MonaRk app.
}
