% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppUtils.R
\name{makeInFlightDf}
\alias{makeInFlightDf}
\title{Load and Organize in-flight data for a route.}
\usage{
makeInFlightDf(route, flightgeoCovar, geomatrix)
}
\arguments{
\item{route}{A string like "SAN SFO" describing the start and end of a route}

\item{flightgeoCovar}{A dataframe of all the flights information only within the given route; "df.RDS"}

\item{geomatrix}{A list of matrices containing 10-second data for lat, long, alt, velocity; "geo.RDS"}
}
\value{
A df with rows representing time, repeated for each flight, columns containing location, weather, delay, and other information shown in tooltips.
}
\description{
This function takes the name of a route, the flight data and geo data for this route,
and returns a df used by leaflet and plotly within the app to display in-flight information.
}
