% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmetParsing.R
\name{parseSigmet}
\alias{parseSigmet}
\title{Parse raw sigmet text into a dataframe}
\usage{
parseSigmet(rawSigmet, day)
}
\arguments{
\item{rawSigmet}{A character vector containing raw sigmet records for one day}

\item{day}{The date in format "YYYY-MM-DD" since records have day but not month or year}
}
\value{
A df with rows as sigmets, and columns as time, location, etc
}
\description{
Loop through each sigmet record to coalesce relevant data like 
time, location, and type. Contains lots of helper functions 
and special cases since there are many styles of sigmet record
}
