% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmetParsing.R
\name{plotSigmet}
\alias{plotSigmet}
\title{Map sigmet data}
\usage{
plotSigmet(sigmetgeo)
}
\arguments{
\item{sigmetgeo}{A df with rows as sigmets, and columns as time, location, etc}
}
\value{
A leaflet map showing sigmet outlines
}
\description{
Pass a parsed sigmetgeo dataframe to see the shapes of the sigmet on a map.
Click a sigmet to see additional data in a tooltip.
}
