% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StormsParsing.R
\name{polygonizeStorms}
\alias{polygonizeStorms}
\title{Create sf dataframe from storm data}
\usage{
polygonizeStorms(storms, distance = 1000)
}
\arguments{
\item{storms}{A df containing cleaned storm event data. Needs BEGIN_LAT, END_LAT, BEGIN_LON, and END_LON columns}
}
\value{
A sf dataframe containing Storm Event information PLUS a geo data column with storm polygons
}
\description{
Pass return from cleanStorms() to this function to create geo data from coordinates
}
