% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatesParsing.R
\name{readOpenSkyStates}
\alias{readOpenSkyStates}
\title{Download, unzip, and read raw OpenSky States Files into R}
\usage{
readOpenSkyStates(
  date,
  range = 0:23,
  download = TRUE,
  fread = TRUE,
  dir = NULL
)
}
\arguments{
\item{date}{One date at a time, in the format "YYYY-MM-DD"}

\item{range}{The hours in the day to be included. Default is all (0:23), but if download fails halfway though, can pick it up (5:23)}

\item{download}{T/F Need to download data from opensky and unzip it first? False if you have already downloaded and want to just read into R}

\item{fread}{T/F Need to read it into R? False if you want to download only}

\item{dir}{Where to download files or where to look for already downloaded files}
}
\value{
"states" which is a dataframe containing a row for each plane at 10 second intervals. This will be fed to additional functions to get data into a usable format.
}
\description{
If states files are not downloaded, use this to download, unzip, and fread into R
If states files are downloaded, use this to unzip, and fread into R by download = F
}
