% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmetParsing.R
\name{readSigmetFiles}
\alias{readSigmetFiles}
\title{Read from a folder containing sigmet txt files}
\usage{
readSigmetFiles(sigmetDataDir)
}
\arguments{
\item{sigmetDataDir}{Location of sigmet TXTs}
}
\value{
A character vector containing sigmet text
}
\description{
Raw sigmet text is stored in separate files for each hour, inside a folder for each day.
For each day that needs to have sigmet data processed, loop though and run this function
to get a character vector, and then parseSigmet() on that vector. Different sources of 
sigmet data contain different record separating characters, reflected in the "split" arguments.
}
