% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCovars.R
\name{thisDelay}
\alias{thisDelay}
\title{Flight delay}
\usage{
thisDelay(flightgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}
}
\value{
A df containing columns with minute delays for weather, NAS, carrier, security, and late aircraft.
}
\description{
Read in delay data. Then loop through US flights to find matching records 
in the delay data by matching origin, destination, callsign, plane, and then time.
}
