% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCovars.R
\name{thisFlightWeather}
\alias{thisFlightWeather}
\title{Flight sigmet overlap}
\usage{
thisFlightWeather(flightgeo, sigmetgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}

\item{sigmetgeo}{A df with rows as sigmets, and columns as time, location, etc}
}
\value{
A T/F vector to show if sigmets overlapped each flight.
}
\description{
Loop through the flights if any sigmet overlapped the flight in space and time
}
