% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCovars.R
\name{thisRegionWeather}
\alias{thisRegionWeather}
\title{Flight region sigmet overlap ID}
\usage{
thisRegionWeather(flightgeo, sigmetgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}

\item{sigmetgeo}{A df with rows as sigmets, and columns as time, location, etc}
}
\value{
A numeric vector to show ids of sigmets within ~1000km of a flight.
}
\description{
Loop through the routes and find the midpoint between the two airports, and transform
the flight data to the UTM zone of the midpoint. Draw a 1000km buffer around midpoint.
Then loop through each flight to see if there was a sigmet overlapping the buffer 
at the same time as the flight.
}
