% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCovars.R
\name{thisStormEvent}
\alias{thisStormEvent}
\title{Storm Events Overlap}
\usage{
thisStormEvent(flightgeo)
}
\arguments{
\item{flightgeo}{A sf dataframe containing flight data and geodata}
}
\value{
A T/F vector to show if a storm event overlapped each flight.
}
\description{
Read in and polygonize storm data. Then loop through US flights to find
if any storm events overlapped the flight in time and space.
}
