% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatesParsing.R
\name{transformStates}
\alias{transformStates}
\title{Parse flight state data}
\usage{
transformStates(states, callsigns = unique(states[, callsign]))
}
\arguments{
\item{states}{A df containing flight data, each row 10 seconds for a single aircraft}

\item{callsigns}{Either pass selected callsigns if only interested in a few, or leave as is to select all callsigns}
}
\value{
a list of "geomatrix" (list of matrices) and "flights" (a df where a flight is a row)
}
\description{
Loop through callsigns and subset to all data with that callsign. Separate into
different flights if more than 30 minutes between records. For each of these flights,
only keep data it is a complete flight: starting and ending below 4500m elevation,
more than 10 minutes long, is not a Cessna or Beech, doesn't have a jump/data error of
more than two decimal degrees, is within 500m elevation and .1 decimal degrees lat and lon 
to nearest airport at start and end of flight.
}
